feat(CheckIn): intergrate changes in develop
