+++ app/build.gradle
@@ -75,6 +75,15 @@ dependencies {
     // logging
     compile 'com.squareup.okhttp3:logging-interceptor:3.9.0'
     testImplementation 'com.squareup.okhttp3:mockwebserver:3.10.0'
+
+    implementation 'com.google.firebase:firebase-core:11.8.0'
+    implementation 'com.google.firebase:firebase-auth:11.8.0'
+    implementation 'com.google.android.gms:play-services-auth:11.8.0'
+    implementation 'com.shobhitpuri.custombuttons:google-signin:1.0.0'
+    implementation 'de.hdodenhof:circleimageview:2.2.0'
+    androidTestCompile 'com.android.support.test:rules:1.0.1'
+    androidTestCompile 'com.android.support.test.uiautomator:uiautomator-v18:2.1.3'
+    androidTestCompile 'com.android.support.test.espresso:espresso-contrib:3.0.1'
 }
 
 apply plugin: 'com.google.gms.google-services'
