+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,21 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
-        android:name=".application.ArtApplication">
-        <activity android:name=".main.MainActivity">
+        android:name=".root.App">
+        <activity android:name=".login.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
+            android:name=".serialentry.SerialEntryActivity"
+            android:label="@string/title_activity_serial_entry"
+            android:theme="@style/AppTheme"/>
+        <activity android:name=".main.MainActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW" />
             </intent-filter>
         </activity>
         <activity android:name=".checkin.CheckInActivity">
