--- app/src/main/java/com/andela/art/injection/component/ApplicationComponent.java
@@ -1,17 +1,0 @@
-package com.andela.art.injection.component;
-
-import android.content.Context;
-
-import com.andela.art.injection.modules.ApplicationModule;
-
-import javax.inject.Singleton;
-
-import dagger.Component;
-import retrofit2.Retrofit;
-
-@Singleton
-@Component(modules = ApplicationModule.class)
-public interface ApplicationComponent {
-    Retrofit exposeRetrofit();
-    Context exposeContext();
-}
