+++ app/src/main/java/com/andela/art/injection/component/MainComponent.java
@@ -3,6 +3,7 @@ package com.andela.art.injection.component;
 import com.andela.art.injection.modules.MainModule;
 import com.andela.art.injection.scope.PerActivity;
 import com.andela.art.main.MainActivity;
+import com.andela.art.root.ApplicationComponent;
 
 import dagger.Component;
