--- app/src/main/java/com/andela/art/injection/modules/ApplicationModule.java
@@ -1,73 +1,0 @@
-package com.andela.art.injection.modules;
-
-import android.content.Context;
-
-import java.util.concurrent.TimeUnit;
-
-import javax.inject.Singleton;
-
-import dagger.Module;
-import dagger.Provides;
-import okhttp3.OkHttpClient;
-import okhttp3.logging.HttpLoggingInterceptor;
-import retrofit2.Retrofit;
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
-import retrofit2.converter.gson.GsonConverterFactory;
-
-@Singleton
-@Module
-public class ApplicationModule {
-
-    private String artBaseUrl;
-    private Context context;
-
-    public ApplicationModule(String artBaseUrl, Context context) {
-        this.artBaseUrl = artBaseUrl;
-        this.context = context;
-    }
-
-    @Singleton
-    @Provides
-    GsonConverterFactory gsonConverterFactory() {
-        return GsonConverterFactory.create();
-    }
-
-    @Singleton
-    @Provides
-    HttpLoggingInterceptor provideHttpLoggingInterceptor() {
-        return new HttpLoggingInterceptor()
-                .setLevel(HttpLoggingInterceptor.Level.BODY);
-    }
-
-    @Singleton
-    @Provides
-    OkHttpClient provideOkHttpClient(HttpLoggingInterceptor loggingInterceptor) {
-        return new OkHttpClient.Builder()
-                .connectTimeout(20, TimeUnit.SECONDS)
-                .addInterceptor(loggingInterceptor)
-                .build();
-    }
-
-    @Singleton
-    @Provides
-    RxJava2CallAdapterFactory provideRxJavaCallAdapterFactory() {
-        return  RxJava2CallAdapterFactory.create();
-    }
-
-    @Singleton
-    @Provides
-    Retrofit provideRetrofit(OkHttpClient client, GsonConverterFactory converterFactory, RxJava2CallAdapterFactory rxJavaCallAdapterFactory) {
-        return new Retrofit.Builder()
-                .baseUrl(artBaseUrl)
-                .addConverterFactory(converterFactory)
-                .addCallAdapterFactory(rxJavaCallAdapterFactory)
-                .client(client)
-                .build();
-    }
-
-    @Provides
-    @Singleton
-    Context provideContext() {
-        return context;
-    }
-}
