+++ app/src/main/java/com/andela/art/serialentry/SerialEntryActivity.java
@@ -3,6 +3,7 @@ package com.andela.art.serialentry;
 import android.content.Intent;
 import android.content.pm.ActivityInfo;
 import android.os.Bundle;
+import android.os.Handler;
 import android.support.annotation.NonNull;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -15,8 +16,9 @@ import android.view.Window;
 import android.view.WindowManager;
 import android.widget.TextView;
 
-import com.andela.art.login.LoginActivity;
 import com.andela.art.R;
+import com.andela.art.login.LoginActivity;
+import com.andela.art.main.MainActivity;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.firebase.auth.FirebaseAuth;
@@ -119,5 +121,17 @@ public class SerialEntryActivity extends AppCompatActivity
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
         return true;
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent i=new Intent(SerialEntryActivity.this, MainActivity.class);
+                startActivity(i);
+            }
+        }, 5000);
     }
 }
